+++ test_binary.py
@@ -326,3 +326,16 @@ def test_writer_prec():
         res = q.binary(prec=4)
         exp = '{v}: expected <{p4}>, got <{res}>.'.format(**locals())
         assert res == p4, exp
+
+
+if __name__ == '__main__':
+    # As a debugging aid allow the tests to be run on their own, outside pytest.
+    # This makes it easier to see and interpret and textual output.
+
+    defined = dict(globals())
+    for k, v in defined.items():
+        if callable(v) and k.startswith('test_'):
+            print()
+            print('Calling:', k)
+            print((len(k)+9)*'=')
+            v()